feat: 즐겨 찾기 삭제 기능 구현
- 인수 테스트 메소드 분리
