+++ src/test/java/nextstep/subway/favorite/acceptance/FavoriteAcceptanceTestMethod.java
@@ -1,0 +1,57 @@
+package nextstep.subway.favorite.acceptance;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import java.util.List;
+import nextstep.subway.favorite.dto.FavoriteRequest;
+import nextstep.subway.favorite.dto.FavoriteResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+public class FavoriteAcceptanceTestMethod {
+
+    public static void 즐겨찾기_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_생성_요청(String token, Long sourceStationId, Long targetStationId) {
+        FavoriteRequest favoriteRequest = new FavoriteRequest(sourceStationId, targetStationId);
+        return RestAssured.given().log().all().
+                auth().oauth2(token).
+                body(favoriteRequest).
+                contentType(MediaType.APPLICATION_JSON_VALUE).
+                when().post("/favorites").
+                then().log().all().extract();
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_목록_조회_요청(String token) {
+        return RestAssured.given().log().all().
+                auth().oauth2(token).
+                when().get("/favorites").
+                then().log().all().extract();
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(String token, ExtractableResponse<Response> response) {
+        String uri = response.header("Location");
+        return RestAssured.given().log().all().
+                auth().oauth2(token).
+                when().delete(uri).
+                then().log().all().extract();
+    }
+
+    public static void 즐겨찾기_목록_삭제됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public static void 즐겨찾기_목록_조회됨(ExtractableResponse<Response> response) {
+        List<FavoriteResponse> favoriteResponses = response.jsonPath().getList(".", FavoriteResponse.class);
+        assertAll(
+                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
+                () -> assertThat(favoriteResponses).hasSize(1)
+        );
+    }
+}
